Sleep briefly before exiting after all missions finished
Glean sends pings asynchronously from another process.
When this runs within a docker container the container will be destroyed
as soon as burnham exits, leaving no time for the uploader to send
pings.
By briefly waiting we let the process handle sending some pings and only
then exit.

Note: the wait time is completely arbitrary. Under normal circumstances
the upload should be finished well within a second, so sleeping a bit
longer is just a precaution.
This will stil result in pings not being sent if they are stuck for
other reasons.